requirements.txt
# Core scientific stack
numpy>=1.26
pandas>=2.1
scipy>=1.11

# Bayesian inference
pymc==5.28.2
arviz==0.23.4

# HTTP client (async)
httpx==0.28.1

# Database
supabase==2.28.2

# Dashboard & visualisation
streamlit==1.55.0
plotly==6.6.0

# ML calibration & metrics
scikit-learn>=1.3

# Claude API (agents)
anthropic==0.49.0

# Web scraping (context agent)
beautifulsoup4==4.13.4

# Progress bars
tqdm==4.67.1

# Environment
python-dotenv==1.2.2

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
